refactor(about): split about view sections
